backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
websockify>=0.11.0
chromadb>=0.4.22
pydantic>=2.6.0
pydantic-settings>=2.1.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
# For Gemini API (direct, not CLI)
google-generativeai>=0.8.0

